feat: add code for Kafka in auth-service

+++ services/auth-service/internal/repository/account_repo.go
@@ -43,7 +43,7 @@ func (r *AccountRepository) GetAccountByUsernamePasswordRole(ctx context.Context
 }
 
 // UpdatePassword update password
-func (r *AccountRepository) UpdatePassword(ctx context.Context, acc *model.Account, newPassword string, pwdVersion int) error {
+func (r *AccountRepository) UpdatePassword(ctx context.Context, acc *model.Account, newPassword string, pwdVersion int64) error {
 	res := r.DB.WithContext(ctx).Model(acc).Select("Password", "PwdVersion").Updates(map[string]interface{}{"Password": newPassword, "PwdVersion": pwdVersion})
 	if res.Error != nil {
 		return res.Error
