+++ services/auth-service/pkg/model/account.go
@@ -3,10 +3,10 @@ package model
 import "gorm.io/gorm"
 
 type Account struct {
-	ID         uint           `gorm:"primaryKey;autoIncrement"`
+	ID         uint64         `gorm:"primaryKey;autoIncrement"`
 	Username   string         `gorm:"unique;not null"`
 	Password   string         `gorm:"not null"`
 	Role       string         `gorm:"not null"`
-	PwdVersion int            `gorm:"not null"`
+	PwdVersion int64          `gorm:"not null"`
 	DeletedAt  gorm.DeletedAt `gorm:"index"`
 }
